# frontend/openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ActiveRequest" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          }
        },
        "required" : [ "active" ],
        "type" : "object"
      },
...
      "AuditEntryView" : {
        "properties" : {
          "action" : {
            "enum" : [ "BOOK", "RESCHEDULE", "MARK", "CANCEL", "CONFIRM", "TRANSFER" ],
            "type" : "string"
          },
          "actorRole" : {
            "enum" : [ "COACH", "STUDENT" ],
            "type" : "string"
          },
          "actorUserId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "method" : {
            "enum" : [ "COACH", "STUDENT", "QR", "LATER" ],
            "type" : "string"
          },
          "newStatus" : {
            "enum" : [ "SCHEDULED", "ATTENDED", "CANCELLED_ON_TIME", "RESCHEDULED", "NO_SHOW", "CANCELLED_BY_COACH" ],
            "type" : "string"
          },
          "occurredAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "previousStatus" : {
            "enum" : [ "SCHEDULED", "ATTENDED", "CANCELLED_ON_TIME", "RESCHEDULED", "NO_SHOW", "CANCELLED_BY_COACH" ],
            "type" : [ "string", "null" ]
          },
          "reason" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "action", "actorRole", "actorUserId", "method", "newStatus", "occurredAt" ],
        "type" : "object"
      },
...
      "ConfirmQrRequest" : {
        "properties" : {
          "token" : {
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "token" ],
        "type" : "object"
      },
      "ConfirmQrResult" : {
        "properties" : {
          "alreadyConfirmed" : {
            "type" : "boolean"
          },
          "attendance" : {
            "$ref" : "#/components/schemas/AttendanceView"
          },
          "markedAttended" : {
            "type" : "boolean"
          }
        },
        "required" : [ "alreadyConfirmed", "attendance", "markedAttended" ],
        "type" : "object"
      },
      "ConfirmResult" : {
        "properties" : {
          "alreadyConfirmed" : {
            "type" : "boolean"
          },
          "attendance" : {
            "$ref" : "#/components/schemas/AttendanceView"
          }
        },
        "required" : [ "alreadyConfirmed", "attendance" ],
        "type" : "object"
      },
...
      "CycleOverview" : {
        "properties" : {
          "cycle" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/CycleSummary"
            }, {
              "type" : "null"
            } ]
          },
          "extension" : {
            "$ref" : "#/components/schemas/ExtensionAvailability"
          },
          "lastPayment" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/PaymentSummary"
            }, {
              "type" : "null"
            } ]
          },
          "payment" : {
            "$ref" : "#/components/schemas/PaymentAvailability"
          },
          "planName" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "extension", "payment" ],
        "type" : "object"
      },
      "CycleSummary" : {
        "properties" : {
          "classesIncluded" : {
            "format" : "int32",
            "type" : "integer"
          },
          "classesLost" : {
            "format" : "int32",
            "type" : "integer"
          },
          "classesRemaining" : {
            "format" : "int32",
            "type" : "integer"
          },
          "classesUsed" : {
            "format" : "int32",
            "type" : "integer"
          },
          "endDate" : {
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "modality" : {
            "enum" : [ "PERSONALIZED", "SEMI_PERSONALIZED" ],
            "type" : "string"
          },
          "originalEndDate" : {
            "format" : "date",
            "type" : "string"
          },
          "pendingMarks" : {
            "format" : "int32",
            "type" : "integer"
          },
          "planId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "startDate" : {
            "format" : "date",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "ACTIVE", "COMPLETED", "EXPIRED" ],
            "type" : "string"
          },
          "studentId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "classesIncluded", "classesLost", "classesRemaining", "classesUsed", "endDate", "id", "modality", "originalEndDate", "pendingMarks", "planId", "startDate", "status", "studentId" ],
        "type" : "object"
      },
...
      "ExtendCycleCommand" : {
        "properties" : {
          "newEndDate" : {
            "format" : "date",
            "type" : "string"
          },
          "reason" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "newEndDate", "reason" ],
        "type" : "object"
      },
      "ExtensionAvailability" : {
        "properties" : {
          "canExtendCycle" : {
            "type" : "boolean"
          },
          "extendFrom" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "extendUntil" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "canExtendCycle" ],
        "type" : "object"
      },
...
      "Header" : {
        "properties" : {
          "cycleEnd" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "cycleStart" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "fullName" : {
            "type" : "string"
          },
          "goal" : {
            "type" : [ "string", "null" ]
          },
          "guardian" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/GuardianView"
            }, {
              "type" : "null"
            } ]
          },
          "studentId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "whatsappPhone" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "fullName", "studentId" ],
        "type" : "object"
      },
...
      "PaymentRegistered" : {
        "properties" : {
          "cycleId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "endDate" : {
            "format" : "date",
            "type" : "string"
          },
          "paymentId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reference" : {
            "type" : [ "string", "null" ]
          },
          "startDate" : {
            "format" : "date",
            "type" : "string"
          }
        },
        "required" : [ "cycleId", "endDate", "paymentId", "startDate" ],
        "type" : "object"
      },
      "PaymentSummary" : {
        "properties" : {
          "amountCop" : {
            "format" : "int64",
            "type" : "integer"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "cycleId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "method" : {
            "enum" : [ "NEQUI", "TRANSFER", "CASH", "OTHER" ],
            "type" : "string"
          },
          "paidOn" : {
            "format" : "date",
            "type" : "string"
          },
          "recordedBy" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reference" : {
            "type" : [ "string", "null" ]
          },
          "studentId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "amountCop", "createdAt", "cycleId", "id", "method", "paidOn", "recordedBy", "studentId" ],
        "type" : "object"
      },
      "PlanInput" : {
        "properties" : {
          "classesIncluded" : {
            "format" : "int32",
            "maximum" : 200,
            "minimum" : 1,
            "type" : "integer"
          },
          "modality" : {
            "enum" : [ "PERSONALIZED", "SEMI_PERSONALIZED" ],
            "type" : "string"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "priceCop" : {
            "format" : "int64",
            "minimum" : 0,
            "type" : "integer"
          }
        },
        "required" : [ "classesIncluded", "modality", "name", "priceCop" ],
        "type" : "object"
      },
      "PlanSummary" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "classesIncluded" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "modality" : {
            "enum" : [ "PERSONALIZED", "SEMI_PERSONALIZED" ],
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "priceCop" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "active", "classesIncluded", "id", "modality", "name", "priceCop" ],
        "type" : "object"
      },
      "QrView" : {
        "properties" : {
          "availableFrom" : {
            "format" : "date-time",
            "type" : "string"
          },
          "availableUntil" : {
            "format" : "date-time",
            "type" : "string"
          },
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          },
          "validForSeconds" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "required" : [ "availableFrom", "availableUntil", "expiresAt", "token", "url", "validForSeconds" ],
        "type" : "object"
      },
...
      "RegisterPaymentCommand" : {
        "properties" : {
          "amountCop" : {
            "exclusiveMinimum" : 0,
            "format" : "int64",
            "type" : "integer"
          },
          "method" : {
            "enum" : [ "NEQUI", "TRANSFER", "CASH", "OTHER" ],
            "type" : "string"
          },
          "overrideModality" : {
            "type" : "boolean"
          },
          "overrideReason" : {
            "type" : "string"
          },
          "paidOn" : {
            "format" : "date",
            "type" : "string"
          },
          "planId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reference" : {
            "maxLength" : 300,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "amountCop", "method", "planId" ],
        "type" : "object"
      },
...
      "SheetView" : {
        "properties" : {
          "classes" : {
            "items" : {
              "$ref" : "#/components/schemas/ClassRow"
            },
            "type" : "array"
          },
          "cycle" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/CycleSummary"
            }, {
              "type" : "null"
            } ]
          },
          "header" : {
            "$ref" : "#/components/schemas/Header"
          },
          "otherEntries" : {
            "items" : {
              "$ref" : "#/components/schemas/ClassRow"
            },
            "type" : "array"
          }
        },
        "required" : [ "classes", "header", "otherEntries" ],
        "type" : "object"
      },
...
      "StudentAccountView" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "hasAccount" : {
            "type" : "boolean"
          },
          "openResetLinkUntil" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "passwordChangedAt" : {
            "format" : "date-time",
            "type" : [ "string", "null" ]
          },
          "passwordChangedBy" : {
            "enum" : [ "SELF", "COACH_LINK" ],
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "active", "hasAccount" ],
        "type" : "object"
      },
...
      "StudentConsentsView" : {
        "properties" : {
          "anonymizationRequested" : {
            "type" : "boolean"
          },
          "audience" : {
            "enum" : [ "ADULT", "GUARDIAN" ],
            "type" : [ "string", "null" ]
          },
          "consents" : {
            "items" : {
              "$ref" : "#/components/schemas/ConsentItemView"
            },
            "type" : "array"
          },
          "studentId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "anonymizationRequested", "consents", "studentId" ],
        "type" : "object"
      },
...
      "StudentProfile" : {
        "properties" : {
          "account" : {
            "$ref" : "#/components/schemas/StudentAccountView"
          },
          "consents" : {
            "$ref" : "#/components/schemas/StudentConsentsView"
          },
          "cycle" : {
            "$ref" : "#/components/schemas/CycleOverview"
          },
          "student" : {
            "$ref" : "#/components/schemas/StudentSummary"
          },
          "upcoming" : {
            "items" : {
              "$ref" : "#/components/schemas/UpcomingClass"
            },
            "type" : "array"
          }
        },
        "required" : [ "account", "consents", "cycle", "student", "upcoming" ],
        "type" : "object"
      },
...
      "StudentSummary" : {
        "properties" : {
          "active" : {
            "type" : "boolean"
          },
          "ageAlert" : {
            "enum" : [ "NONE", "TURNS_ADULT_SOON", "TURNED_ADULT_NEEDS_AUTHORIZATION" ],
            "type" : [ "string", "null" ]
          },
          "anonymizationRequested" : {
            "type" : "boolean"
          },
          "audience" : {
            "enum" : [ "ADULT", "GUARDIAN" ],
            "type" : [ "string", "null" ]
          },
          "birthDate" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "daysUntilAdult" : {
            "format" : "int64",
            "type" : "integer"
          },
          "email" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "goal" : {
            "type" : [ "string", "null" ]
          },
          "guardian" : {
            "oneOf" : [ {
              "$ref" : "#/components/schemas/GuardianView"
            }, {
              "type" : "null"
            } ]
          },
          "hasAccount" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "minor" : {
            "type" : "boolean"
          },
          "turnsAdultOn" : {
            "format" : "date",
            "type" : [ "string", "null" ]
          },
          "whatsappPhone" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "active", "anonymizationRequested", "daysUntilAdult", "email", "fullName", "hasAccount", "id", "minor" ],
        "type" : "object"
      },
...
      "UpcomingClass" : {
        "properties" : {
          "attendanceId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "endsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "eventId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "modality" : {
            "enum" : [ "PERSONALIZED", "SEMI_PERSONALIZED" ],
            "type" : "string"
          },
          "startsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "today" : {
            "type" : "boolean"
          }
        },
        "required" : [ "attendanceId", "endsAt", "eventId", "modality", "startsAt", "today" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/coach/attendances/{id}/audit" : {
      "get" : {
        "operationId" : "audit",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AuditEntryView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "attendance-controller" ]
      }
    },
...
    "/api/coach/cycles/{cycleId}/extend" : {
      "post" : {
        "operationId" : "extend",
        "parameters" : [ {
          "in" : "path",
          "name" : "cycleId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExtendCycleCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CycleSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "billing-controller" ]
      }
    },
...
    "/api/coach/events/{id}/qr" : {
      "get" : {
        "operationId" : "qr",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QrView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "attendance-controller" ]
      }
    },
    "/api/coach/payments" : {
      "get" : {
        "operationId" : "payments",
        "parameters" : [ {
          "in" : "query",
          "name" : "studentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PaymentSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "billing-controller" ]
      }
    },
    "/api/coach/plans" : {
      "get" : {
        "operationId" : "listPlans",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PlanSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "billing-controller" ]
      },
      "post" : {
        "operationId" : "createPlan",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PlanInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlanSummary"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "billing-controller" ]
      }
    },
    "/api/coach/plans/{id}" : {
      "put" : {
        "operationId" : "updatePlan",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PlanInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlanSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "billing-controller" ]
      }
    },
    "/api/coach/plans/{id}/active" : {
      "patch" : {
        "operationId" : "setPlanActive",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ActiveRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PlanSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "billing-controller" ]
      }
    },
...
    "/api/coach/students/{id}/profile" : {
      "get" : {
        "operationId" : "profile",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StudentProfile"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "attendance-controller" ]
      }
    },
    "/api/coach/students/{id}/sheet" : {
      "get" : {
        "operationId" : "sheet",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "cycleId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SheetView"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "attendance-controller" ]
      }
    },
...
    "/api/coach/students/{studentId}/cycles" : {
      "get" : {
        "operationId" : "cycles",
        "parameters" : [ {
          "in" : "path",
          "name" : "studentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CycleSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "billing-controller" ]
      }
    },
    "/api/coach/students/{studentId}/cycles/active" : {
      "get" : {
        "operationId" : "activeCycle",
        "parameters" : [ {
          "in" : "path",
          "name" : "studentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CycleSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "billing-controller" ]
      }
    },
    "/api/coach/students/{studentId}/payments" : {
      "post" : {
        "operationId" : "registerPayment",
        "parameters" : [ {
          "in" : "path",
          "name" : "studentId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterPaymentCommand"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaymentRegistered"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "billing-controller" ]
      }
    },
...
    "/api/student/attendances/confirm-qr" : {
      "post" : {
        "operationId" : "confirmQr",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConfirmQrRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConfirmQrResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "attendance-controller" ]
      }
    },
    "/api/student/attendances/{id}/confirm" : {
      "post" : {
        "operationId" : "confirm",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConfirmResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "attendance-controller" ]
      }
    },
...
  }
}